+++ src/com/vlfom/steplearn/robot/Tib.java
@@ -2,9 +2,9 @@ package com.vlfom.steplearn.robot;
 
 public class Tib extends BodyPart {
     public int length;
-    public double angle;
+    public int angle;
 
-    public Tib(int length, double angle, int weight) {
+    public Tib(int length, int angle, int weight) {
         this.length = length;
         this.angle = angle;
         this.weight = weight;
